Bug 1874684 - Part 9: Prefer integer types for DateDuration. r=mgaudet
Similar to previous part, prefer `int64_t` rather than `double` for `DateDuration`.

Differential Revision: https://phabricator.services.mozilla.com/D198543

UltraBlame original commit: 55bd9a447858c54a1c277a3a72f4f36ed324ad39

+++ js/src/builtin/temporal/PlainYearMonth.cpp
@@ -508,7 +508,7 @@ static bool DifferenceTemporalPlainYearMonth(JSContext* cx,
   }
 
   
-  auto dateDuration = DateDuration{until.years, until.months};
+  auto dateDuration = DateDuration{int64_t(until.years), int64_t(until.months)};
 
   
   if (settings.smallestUnit != TemporalUnit::Month ||
